+++ tests/QueryType/Stream/QueryTest.php
@@ -3,7 +3,10 @@
 namespace Solarium\Tests\QueryType\Stream;
 
 use PHPUnit\Framework\TestCase;
+use Solarium\Core\Client\Client;
 use Solarium\QueryType\Stream\Query;
+use Solarium\QueryType\Stream\RequestBuilder;
+use Solarium\QueryType\Stream\ResponseParser;
 
 class QueryTest extends TestCase
 {
@@ -24,9 +27,30 @@ class QueryTest extends TestCase
         $this->assertSame('e1', $q->getExpression());
     }
 
+    public function testGetType()
+    {
+        $this->assertSame(Client::QUERY_STREAM, $this->query->getType());
+    }
+
+    public function testGetRequestBuilder()
+    {
+        $this->assertInstanceOf(RequestBuilder::class, $this->query->getRequestBuilder());
+    }
+
+    public function testGetResponseParser()
+    {
+        $this->assertInstanceOf(ResponseParser::class, $this->query->getResponseParser());
+    }
+
     public function testSetAndGetExpression()
     {
         $this->query->setExpression('testexpression');
         $this->assertSame('testexpression', $this->query->getExpression());
+    }
+
+    public function testSetAndGetDocumentClass()
+    {
+        $this->query->setDocumentClass('testdocumentclass');
+        $this->assertSame('testdocumentclass', $this->query->getDocumentClass());
     }
 }